chore: move exceptions to exception folder
