+++ src/main/java/com/aws/greengrass/deployment/DefaultDeploymentTask.java
@@ -17,9 +17,9 @@ import com.aws.greengrass.deployment.model.Deployment;
 import com.aws.greengrass.deployment.model.DeploymentDocument;
 import com.aws.greengrass.deployment.model.DeploymentResult;
 import com.aws.greengrass.deployment.model.DeploymentTask;
-import com.aws.greengrass.deployment.templating.RecipeTransformerException;
+import com.aws.greengrass.deployment.templating.exceptions.RecipeTransformerException;
 import com.aws.greengrass.deployment.templating.TemplateEngine;
-import com.aws.greengrass.deployment.templating.TemplateExecutionException;
+import com.aws.greengrass.deployment.templating.exceptions.TemplateExecutionException;
 import com.aws.greengrass.logging.api.Logger;
 import com.aws.greengrass.util.Coerce;
 import com.vdurmont.semver4j.Semver;
