+++ src/main/java/com/aws/greengrass/deployment/templating/TemplateEngine.java
@@ -11,7 +11,10 @@ import com.aws.greengrass.componentmanager.ComponentStore;
 import com.aws.greengrass.componentmanager.exceptions.PackageLoadingException;
 import com.aws.greengrass.componentmanager.models.ComponentIdentifier;
 import com.aws.greengrass.deployment.templating.exceptions.IllegalTemplateDependencyException;
+import com.aws.greengrass.deployment.templating.exceptions.IllegalTransformerException;
 import com.aws.greengrass.deployment.templating.exceptions.MultipleTemplateDependencyException;
+import com.aws.greengrass.deployment.templating.exceptions.RecipeTransformerException;
+import com.aws.greengrass.deployment.templating.exceptions.TemplateExecutionException;
 import com.aws.greengrass.util.NucleusPaths;
 import com.aws.greengrass.util.Pair;
 import com.aws.greengrass.util.Utils;
