+++ src/main/java/com/aws/greengrass/deployment/templating/exceptions/IllegalTemplateParameterException.java
@@ -5,8 +5,6 @@
 
 package com.aws.greengrass.deployment.templating.exceptions;
 
-import com.aws.greengrass.deployment.templating.TemplateParameterException;
-
 public class IllegalTemplateParameterException extends TemplateParameterException {
     private static final long serialVersionUID = 6638067823749416076L;
