+++ src/main/java/com/aws/greengrass/deployment/templating/exceptions/MultipleTemplateDependencyException.java
@@ -5,8 +5,6 @@
 
 package com.aws.greengrass.deployment.templating.exceptions;
 
-import com.aws.greengrass.deployment.templating.TemplateExecutionException;
-
 public class MultipleTemplateDependencyException extends TemplateExecutionException {
     private static final long serialVersionUID = 4897628950643037974L;
