+++ src/main/java/com/aws/greengrass/deployment/templating/exceptions/RecipeTransformerException.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.deployment.templating;
+package com.aws.greengrass.deployment.templating.exceptions;
 
 public class RecipeTransformerException extends Exception {
     private static final long serialVersionUID = -5983084851017096286L;
