+++ src/main/java/com/aws/greengrass/deployment/templating/exceptions/TemplateExecutionException.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.deployment.templating;
+package com.aws.greengrass.deployment.templating.exceptions;
 
 public class TemplateExecutionException extends Exception {
     private static final long serialVersionUID = 8270613415461482760L;
